api/requirements.txt
Flask-SQLAlchemy
Flask-JWT-Extended
pymysql
python-dotenv
flask-uploads
flask-cors
Flask>=2.0.3
Werkzeug>=2.0.3
gunicorn==20.1.0
requests
google-auth
flask_mail


